feat: add pointers
Adds pointers which upon creation will point to some existing value
in memory. Pointers can be dereferenced to read the value from memory.
Pointers to struct fields can also be created and dereferenced.
A value in memory can also be changed through a pointer but only if
the pointer is declared `mut`. A `mut` pointer can only be created
if the memory it points to was also declared as `mut` previously.